+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added Toky Night theme
 - Added Nord Light/Dark themes
 - Added Pop! Light/Dark themes
+- Added Yaru Light/Dark themes
 - Added Adwaita Light/Dark themes
 - Added Breeze Light/Dark themes